prueba de test fallido

+++ test/fiuba/algo3/tp2/integracion/entrega1/HerramientaTest.java
@@ -15,6 +15,6 @@ public class HerramientaTest {
 		Herramienta hacha = new Hacha();
 		
 		assertEquals(new Integer(100), hacha.getDurabilidad());
-		assertEquals(new Integer(2), hacha.getFuerza());
+		assertEquals(new Integer(3), hacha.getFuerza());
 	}
 }
